Version_1.1/DRonA_MLSynergy_requirements.txt
pandas
biopython
numpy
sklearn
holoviews
bokeh